+++ CHANGELOG.md
@@ -127,6 +127,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Items sort : The items in a list are now sorted alphabeti
 - New account : A new account can now create a wishlist and share it
 - Authentication : user now have error messages on authentication/registration
+- Google auth : temporarily disable google auth button
 - Color : colors are not pastel anymore
 
 ### Security